requirements.txt
Django==2.2.1
django-cors-headers==3.0.1
djangorestframework==3.9.3
pytz==2019.1
sqlparse==0.3.0
